Remove code for generating static vcs.
