+++ .github/workflows/main.yml
@@ -41,12 +41,6 @@ jobs:
           node-version: ${{ matrix.node-version }}
       - name: Install npm dependencies
         run: npm install --legacy-peer-deps
-      - name: Clean VCs
-        run: rm -f ./static-vcs/*.json
-      - name: Generate VCs
-        env:
-          CLIENT_SECRET_DB: ${{ secrets.CLIENT_SECRET_DB }}
-        run: npm run generate-vcs
       - name: Run test with Node.js ${{ matrix.node-version }}
         env:
           CLIENT_SECRET_DB: ${{ secrets.CLIENT_SECRET_DB }}
