--- static-vcs/validVcForRevocation.json
@@ -1,30 +1,0 @@
-{
-  "@context": [
-    "https://www.w3.org/2018/credentials/v1",
-    "https://w3id.org/vc/status-list/2021/v1",
-    "https://w3id.org/security/suites/ed25519-2020/v1"
-  ],
-  "type": [
-    "VerifiableCredential"
-  ],
-  "credentialSubject": {
-    "id": "did:key:z6MkhTNL7i2etLerDK8Acz5t528giE5KA4p75T6ka1E1D74r"
-  },
-  "id": "urn:uuid:3c265c65-98d4-497c-bc68-f5d278e7ed73",
-  "issuanceDate": "2022-06-29T19:45:10Z",
-  "issuer": "did:key:z6MksxoKCsNUXQLNjrNKaMjq4GDgc7djeqTziY68z2qk5jKC",
-  "credentialStatus": {
-    "id": "https://localhost:39443/issuers/z19saGo3QDCewbDXMubPErpbq/slcs/z19yNB4mGSWHNq6bEHP4d5Ziw#39811",
-    "type": "StatusList2021Entry",
-    "statusListCredential": "https://localhost:39443/issuers/z19saGo3QDCewbDXMubPErpbq/slcs/z19yNB4mGSWHNq6bEHP4d5Ziw",
-    "statusListIndex": "39811",
-    "statusPurpose": "revocation"
-  },
-  "proof": {
-    "type": "Ed25519Signature2020",
-    "created": "2022-06-29T19:45:10Z",
-    "verificationMethod": "did:key:z6MksxoKCsNUXQLNjrNKaMjq4GDgc7djeqTziY68z2qk5jKC#z6MksxoKCsNUXQLNjrNKaMjq4GDgc7djeqTziY68z2qk5jKC",
-    "proofPurpose": "assertionMethod",
-    "proofValue": "z3sF6vh1CGuBk9JPd8kaRSPQ9seWHfxBFS4qFcYWtrZ48khh5dT7xWXzzTHLnj6pZ25TsPEJWJfH6hJsnoNzEQMDr"
-  }
-}
