--- static-vcs/validVcForSuspension.json
@@ -1,30 +1,0 @@
-{
-  "@context": [
-    "https://www.w3.org/2018/credentials/v1",
-    "https://w3id.org/vc/status-list/2021/v1",
-    "https://w3id.org/security/suites/ed25519-2020/v1"
-  ],
-  "type": [
-    "VerifiableCredential"
-  ],
-  "credentialSubject": {
-    "id": "did:key:z6MkhTNL7i2etLerDK8Acz5t528giE5KA4p75T6ka1E1D74r"
-  },
-  "id": "urn:uuid:d0db6513-1b91-4166-82d3-2183d1110353",
-  "issuanceDate": "2022-06-29T19:45:10Z",
-  "issuer": "did:key:z6Mktq3BbtU74LtiTw1fzwuE3KCP7f4XeUNN1LHoXTCEzdCN",
-  "credentialStatus": {
-    "id": "https://localhost:39443/issuers/z19u3vMn3YdmoH4QkTxzypi1N/slcs/z19wSPK32CB4n9b9Aiujm7WVE#129411",
-    "type": "StatusList2021Entry",
-    "statusListCredential": "https://localhost:39443/issuers/z19u3vMn3YdmoH4QkTxzypi1N/slcs/z19wSPK32CB4n9b9Aiujm7WVE",
-    "statusListIndex": "129411",
-    "statusPurpose": "suspension"
-  },
-  "proof": {
-    "type": "Ed25519Signature2020",
-    "created": "2022-06-29T19:45:10Z",
-    "verificationMethod": "did:key:z6Mktq3BbtU74LtiTw1fzwuE3KCP7f4XeUNN1LHoXTCEzdCN#z6Mktq3BbtU74LtiTw1fzwuE3KCP7f4XeUNN1LHoXTCEzdCN",
-    "proofPurpose": "assertionMethod",
-    "proofValue": "z3yAQivxy22hZQJQQ7jQS3sFy24A39QC7AWoTwYj2vwF9JKtCzJFe932BcX3zoyC77pqxxJACMk5D4ZxVrLS3krsN"
-  }
-}
